requirements/dev_local.txt
# PIP requirements for AWX development/build environment (using only local
packages).  Install using "pip --no-index -r dev_local.txt".

distribute-0.6.45.tar.gz

Django-1.5.1.tar.gz

# Needed by python-dateutil, django-extensions:
six-1.3.0.tar.gz

# Needed by kombu:
amqp-1.0.11.tar.gz
anyjson-0.3.3.tar.gz

# Needed by celery:
billiard-2.7.3.28.tar.gz
kombu-2.5.10.tar.gz
python-dateutil-2.1.tar.gz

# Needed by django-celery:
celery-3.0.19.tar.gz
pytz-2013b.tar.gz

# Remaining dev/prod packages:
django-celery-3.0.17.tar.gz
django-extensions-1.1.1.tar.gz
django-jsonfield-0.9.10.tar.gz
django-taggit-0.10a1.tar.gz
djangorestframework-2.3.5.tar.gz
Markdown-2.3.1.tar.gz
pexpect-2.4.tar.gz
requests-1.2.3.tar.gz
South-0.8.1.tar.gz

# Remaining dev-only packages:
django-debug-toolbar-0.9.4.tar.gz
django-devserver-0.6.2.tar.gz
ipython-0.13.2.zip

# You may also need to install the following extra packages using the OS
# package manager, or pip if you're running inside a virtualenv.
# - ansible (via yum, pip or source checkout)
# - psycopg2 (via "yum install python-psycopg2")
# - coverage-3.6.tar.gz (if you want to check test coverage; the default
#   python-coverage package is old.)
# - readline-6.2.4.1.tar.gz (for using the ipython interactive shell)
